debug: Store disabled log levels in a hashtable
Previously, disabled log levels were stored in a mutable association list.
Each call to a logging function searched the list to find out whether
the log should be emitted.    Currently almost all logs are turned on
by default, so the list is empty or nearly empty and searching it does
not take long.   When we start to turn more logging off by default,
the O(n) cost of searching this list will become much more significant.
This change replaces the list with a hashtable which can be searched in
O(1) time.

Signed-off-by: Euan Harris <euan.harris@citrix.com>