Update OASIS, don't check in files it generates
It is necessary to update the version of the _oasis file, because
OASISFormat 0.2 support has been dropped from the new 0.4.7 OASIS
version: https://forge.ocamlcore.org/frs/shownotes.php?release_id=1223

I've also removed the files generated by OASIS by running "oasis
setup-clean -replace-sections -remove", to avoid keeping auto-generated
files under version control. From now on we will have to run "oasis
setup" to generate the main part of the Makefile and other required
build files, this will require an update to the spec files.

I tried to preserve as much of the original configuration as possible.
The flags defined in the Makefile will be passed to the corresponding
OASIS call, and it seems that OASIS passes these extra arguments to the
program it invokes. In this case, "-j 4" will be passed to ocamlbuild to
specify the number of parallel jobs, and the -verbose flag will be
passed to the test executable, which uses OUnit.

Also update the .gitignore:
- ignore files generated by OASIS
- add a leading "/" to files in the root directory
- append a trailing "/" to directories

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>