Update INSTALL.md: include "oasis setup" step
Now there is an additional step required to build and install the
project when using xs-shipyard, one has to run "oasis setup" to generate
the build files, including the Makefile.

It's important to highlight this step, as it is not mentioned in the
README of xs-shipyard - those instructions assume that a Makefile is
already present.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>