CP-28659: add nvram parameter
This will be used initially for UEFI guests to store:
`EFI-variables`, `EFI-variables-on-boot` and `EFI-variables-backend`.
However it is meant to be extensible in the future, hence a string map.

Unfortunately the type has to be `list option` instead of simply `list`,
otherwise `xenopsd` fails on startup to import existing VMs which did
not have this field.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>