Unwrapped xapi-idl's Syslog module conflicts with Core's
The latest version of Core has moved Core.Unix.Syslog to a separate
Syslog library, making it impossible to link both core and xapi-idl into same binary
(as xapi-storage-script would attempt to do).

Make available both a wrapped and unwrapped library, this alone is sufficient to make
xapi-storage-script able to link again, because it will use the wrapped version for calls
internal to xapi-idl (it never accesses xapi-idl's syslog module).

It is a bit odd to have Xapi_idl.Xcp_service as the new fully wrapped name,
so propose to use Xapi_idl.Service instead, etc.

We can make the actual updates to remove the deprecated usages sometime in the future.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>