CA-326244: avoid deadlock due to Logs library being called from hooks
We could deadlock in this module if we try to log from a finalizer, exception hook, etc.
that got invoked while we were already holding the loglevel_m mutex.
This could happen in libraries that use the Logs library, because now
logs from the Log library go through this Debug module.

So now I have changed the Debug module so that logs from the Logs library
do not need a mutex - now we log all logs from the Logs library at levels
warning or above, regardless of what is enabled in the Debug module.

When we implement CP-28733 to use the Logs library under the hood in the Debug
module and make it a wrapper around the Logs library, we can probably control
log levels of libraries using Logs again through the Debug module.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>